langgraph-chroma-api/pyproject.toml
[project]
name = "Agentic-OCR"
version = "0.1.0"
description = "An Agentic OCR project by Dilavar Afreed"
authors = []
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi (>=0.120.1,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "openai (>=2.6.1,<3.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "chromadb (>=0.4.0)",
    "langchain (>=1.2.9,<2.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pypdf (>=4.0.0,<5.0.0)",
    "python-multipart (>=0.0.6,<0.1.0)",
    "torch (>=2.0.0)",
    "transformers (>=4.40.0)",
    "pillow (>=10.0.0)",
    "pymupdf (>=1.24.0)"
]

[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "run:start"
